[M2-6040][M2-5711][M2-5768]: Take Now flow (#1646)
This PR introduces several new things to kick off the "Take Now" flow. They are: the `TakeNowModal` component, the `useTakeNowModal` hook, and the `enableMultiInformantTakeNow` feature flag. In this PR, I also enable launching of the modal from the Applet Activities Overview and Participant Detail Activities Tab screens.

#### The `TakeNowModal` component and `useTakeNowModal` hook

Use of the modal is restricted through the use of the use of the `useTakeNowModal` hook. I did this so that the consumer of the modal could control its presence by simply calling the `openTakeNowModal` function, which is returned by the hook. This is an intentional departure from the way modals are currently implemented, which has been to expose control of the `open` state to the parent.

The modal component itself only accepts an `onClose` prop, so that the caller may be notified of when the modal is closed. Otherwise, the modal controls its own open state by unmounting itself

The `openTakeNowModal` function accepts an activity, and optional default values for the subject and participant fields.

#### The `enableMultiInformantTakeNow` feature flag

This feature flag controls whether the take now feature is exposed. By default, the take now menu option is presented to users with the roles of `Admin`, `Manager`, and `Coordinator`. The option will be generally be missing for any other user type, and it will be missing for all users when this flag is turned off.