fix: Pin button fixes (M2-6843) (#1777)
* fix: add keyExtractor support to tables

It's bad practice to striclty use array index for React key props, which
is what Table and DashboardTable were using, causing residual styling
to be applied to the wrong element when items shift. Added support for
`keyExtractor`s to both Table and Dashboard table so that row IDs can be
used as keys instead. Required adding support for hidden (no content)
columns.

* fix: Pin styling on focus

Make Pin styling better represent its active state when focused, and
also make its focus/hover state clearer.

* test: update DashboardTable & Table tests

* fix: sorting issue when pinning/unpinning

In `useTable` hook, `handleReload` was not using the most current value
of the `params` variable due to a closure reference error (which I still
don't fully understand), but the suggested fix of wrapping it in a ref
solved it. As a bonus, doing so also reduced redundant rerenders.