feat(mi): Soft-lock Admin App during Take Now (M2-7109) (#1851)
* feat: add soft-lock warning banner

* feat: add soft-lock state data to Auth

Soft-locking requires preserving a portion of the user's session state,
so as to make resuming their session as frictionless as possible. This
includes the URL they were on when logging out, their active workspace,
and their email address (to pre-populate the login form).

* feat: update `useLogout` to store soft-lock data

Also update `TakeNowModal` to take advantage of the hook's new prop.

Note that because the signature of the function returned by `useLogout`
has changed, we had to wrap it in a closure in `LockForm` so as to keep
TypeScript happy.

Also note that `LockForm` is a legacy component that is not currently in
use, and may not be used in the future depending on whether this
implementation of soft-lock proves successful as an alternative
approach. (`LockForm` was created to be activated after 15 minutes of
inactivity, but never fully implemented.)

* feat: make use of soft-lock state on Login screen

Displays banner explaining why you were locked out, prepopulates login
form, focuses password field, and restores nav state on successful login
as same user.

Soft lock state is cleared after restoring nav state, or if navigating
away from the login screen.

Password field focus required updating `InputController` to properly
assign ref so that we can make use of the `setFocus` function returned
by `useForm`.

* test: add soft-lock state test, fix other tests