refactor: Remove release 1 feature flags (M2-6906) (#1878)
This PR removes all references to the `enableMultiInformant` and `enableMultiInformantTakeNow` feature flags. I've also taken the liberty of removing some unused components and related functions. There are probably others that could be removed, and I'd be happy to oblige if you have suggestions.

As part of this ticket I also made two related changes:
- The View Users menu item on the dashboard now points to the participants tab, instead of the applet overview tab
- The applet breadcrumb will now navigate to the applet overview tab, instead of the participants tab