feat: Configure Report To Select Scores or Raw Scores (CAMHI) (M2-7738) (#1935)
* Add support for a tooltip icon with a tooltip in the RadioGroupController

* Add radio buttons with hardcoded text

* Update ScoreReport type to include scoreType and linkedSubscaleName

* WIP update to score type selection logic

* Set selected items based on subscale

* Update score range calculation to account for linked subscale

* Exclude subscales with no activity items

* Pass the `onChange` to the `onCustomChange` in the InputController

This allows callers who pass a custom onChange function to this component to receive and optionally invoke the original react-hook-form onChange. This removes the need to invoke `setValue` (or equivalent) manually, which the docs say you should avoid

* Update report score linked subscale name when it is edited on the subscales configuration screen

* Fix type name

* Require linked subscale when report score type is score

* Update the logic that accounts for updates on the subscales configuration screen

This block of code now also manually removes the linked subscale name if it can't be used to find a valid subscale

* Remove the link to a report score if a subscale or its lookup table is deleted

* Translate strings

* Fix broken tests

* Move key prop to parent

I wrapped this component and forgot to move the key up a level

* Fix MUI select warning

Apparently MUI doesn't like a default value of `undefined` on the select, so the empty string works while still showing the placeholder

* Set score type back to raw score if there are no eligible subscales

* Hide the radio group if there are no eligible subscales

* Fix broken test

* Abstract field names

* Rename scoreType to scoringType

* Rename linkedSubscaleName to subscaleName

* Change `rawScore` to `raw_score`

* Make the `scoringType` field nullable

If there are other activities in the applet with existing report scores that don't have a scoringType, they will prevent any updates to the applet. So we must make this field nullable

* Remove the noisy MUI controlled field warning

* Handle the case where the score is undefined

🤷

* Filter out ineligible subscale items

Subscales can contain other subscales, but we shouldn't include them in the report

* Account for existing applets with reports but no subscales

* Account for reports that were saved without an explicit scoring type

* Reset report scores to raw_score if there are no more eligible subscales

* Remove linked subscales from reports when they no longer contain non-subscale items

* Fix score range calculation

* Ignore score ranges

The lookup table parsing allows ranges in the score column, but this is invalid as scores can only be numbers. So I'm updating this code to ignore those until we can update the parsing code