fix: By Participant list for limited respondents (M2-8092) (#1974)
* fix: By Participant list for limited respondents

Treat limited account as possible valid respondents, since they could
have had submissions done on their behalf via Take Now.

* fix: update empty state copy for limited accounts

* fix: assign logic for limited respondents

Ensured the `isLimitedRespondent` is only true if a respondent subject
is even present. This ensures assign/unassign menu items follow correct
logic on the About Participant tab.

Also, for the By Participant tab, checks first if the respondent is a
limited account before hiding the Assign menu item and showing the
Unassign menu item, so that when an activity or flow is set to
auto-assign, the user isn't misled into thinking that auto-assigned
activities/flows could have been assigned to the limited account as the
respondent.