chore: Remove managers query from ParticipantDropdown (M2-8462) (#2011)
The `useParticipantDropdown` hook previously tried to eagerly fetch the list of team members for filtering the list of respondents, which doesn’t work for users who don’t have permission to consume the workspace managers endpoint (`/workspaces/{owner_id}/applets/{applet_id}/managers`). This PR removes the managers query from the `useParticipantDropdown` hook, and instead makes use of the changes in https://github.com/ChildMindInstitute/mindlogger-backend-refactor/pull/1705 to do the filtering of respondents.

The consequence of this change is that the assign activity drawer is no longer broken for coordinator users. It also prevents the hook from infinitely re-rendering offscreen for other users who don't have permission to consume this endpoint (e.g. reviewers)

I've also fixed the permission set for coordinators, allowing them to pre-populate the assign activity drawer with a respondent or target subject.