feat: New score calculation behaviour for skipped items (M2-8684) (#2027)
* feat: handle skipped items in calculations

When feature flag is enabled, skipped items are not counted in the
subscales calculations or activity total scores. If all items were
skipped and subscale score is `null`, then subscale score is simply
omitted from the report. Same is true for the activity total score.

* feat: handle `null` scores in data export

* fix: bug exporting skipped MultipleSelectionPerRow

* test: fix existing tests

* test: add skipped items tests & fix `calcScores`

In adding unit tests, some kept failing and it helped me discover
a bug in how I was calling `calcScores` from `getSubscales` (a
refactoring relic 🤦🏻‍♂️). While doing so, I opted to better organize
and logic and typing in `calcScores`, which was needlessly obtuse
and cryptic. In turn this may have fixed some further edge-case bugs.

* fix: handle possible `null` subscale value