fix: EMA Fast Follows (M2-9162) (M2-9158) (#2077)
This PR fixes a couple EMA issues. 

For M2-9162, the schedule history file currently does not accurately reflect the behavior of the MindLogger application where participants revert to the default schedule after their individual schedule has been deleted. Specifically this ticket describes the scenario where there are multiple versions of the individual schedule before it is deleted.

For M2-9158, this PR fixes the integration of the new RTK Query method of fetching respondents. The request was not sending an explicit *limit* query parameter, and it seems the API has a default limit of 10. This in turn caused the method that calculates the total number of pages to fail because it assumes that the limit is set to 250 (The ceiling of 10/250 = 1 page). Therefore all applets would export schedule history for no more than 10 respondents. My fix for this makes the assumption of 250 explicit to the caller by providing a parameter for it